transport: move the path shortening functions to unix_unix.go
